fix - Added manual approval for workflow start on a PR
Issue #, if available: n/a

Description of changes: updated build.yml with check for github owner and if not a manual approval is needed to begin workflow

Testing done: created dummy github to test out and verified correct implementation

Merge Checklist
Put an x in the boxes that apply. You can also fill these out after creating the PR. If you're unsure about any of them, don't hesitate to ask. We're here to help! This is simply a reminder of what we are going to look for before merging your pull request.

General
[x] I have read the CONTRIBUTING doc
[x] I certify that the changes I am introducing will be backward compatible
 [x]I used the commit message format described in CONTRIBUTING doc
Tests
[ ] I have added tests that prove my fix is effective or that my feature works (if appropriate)
[ ]  I have added unit and/or integration tests as appropriate to ensure backward compatibility of the changes
[x]  I have verified all code in this commit are well formatted
By submitting this pull request, I confirm that my contribution is made under the terms of the Apache 2.0 license.